updating primary keys. changed to serial

+++ src/main/resources/db/migrations/changes/18112019_create_table_schedule.sql
@@ -1,6 +1,6 @@
 create table schedule
 (
-    id         bigint primary key not null,
+    id         serial primary key not null,
     begin_time time               not null,
     end_time   time               not null
 )
